chore: Migrate to 4.0 "nested" frame implementation and change core modules dep to "next"
