+++ sdk/package.json
@@ -18,10 +18,9 @@
     }
   },
   "dependencies": {
-    "nativescript-fresco": "^3.0.2",
     "nativescript-pro-ui": "next",
     "nativescript-theme-core": "~1.0.4",
-    "tns-core-modules": "^3.4.0"
+    "tns-core-modules": "next"
   },
   "devDependencies": {
     "babel-traverse": "6.9.0",
